requirements.txt
Flask==3.1.3
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1

python-dotenv==1.2.1

google-generativeai==0.8.5
protobuf==3.20.3

pdfplumber==0.11.7
reportlab==4.4.4
